+++ src/renderer/command-center/macros-ui/create-macro.tsx
@@ -1,8 +1,4 @@
 import { useState } from "react";
-import { TriggerType } from "common/types";
-import { HotkeyTrigger } from "@macros/hotkey-trigger";
-import { MouseTrigger } from "@macros/mouse-trigger";
-import { ManualTrigger } from "@macros/manual-trigger";
 import { WorldEventTrigger } from "@macros/world-event-trigger";
 import { useMacroStore } from "./use-macros-store";
 
@@ -12,10 +8,6 @@ export const CreateMacro = ({
   onCreated: (id: string) => void;
 }) => {
   const [name, setName] = useState("");
-  const [triggerType, setTriggerType] = useState<TriggerType>(
-    TriggerType.Hotkey
-  );
-
   const { createMacro } = useMacroStore();
 
   return (
@@ -42,42 +34,12 @@ export const CreateMacro = ({
           />
         </label>
 
-        <label>
-          Trigger Type:
-          <select
-            onChange={(e) => {
-              const type =
-                TriggerType[e.target.value as keyof typeof TriggerType];
-              setTriggerType(type);
-            }}
-            value={triggerType}
-          >
-            {Object.keys(TriggerType).map((key) => (
-              <option key={key} value={key}>
-                {key}
-              </option>
-            ))}
-          </select>
-        </label>
-
         <button
           onClick={async () => {
             if (name.trim() === "") {
               return;
             }
-            let trigger:
-              | ManualTrigger
-              | HotkeyTrigger
-              | MouseTrigger
-              | WorldEventTrigger = new ManualTrigger();
-            if (triggerType === TriggerType.Hotkey) {
-              trigger = new HotkeyTrigger();
-            } else if (triggerType === TriggerType.Mouse) {
-              trigger = new MouseTrigger();
-            } else if (triggerType === TriggerType.WorldEvent) {
-              trigger = new WorldEventTrigger();
-            }
-            const id = await createMacro(name, trigger);
+            const id = await createMacro(name, new WorldEventTrigger());
             onCreated(id);
             setName("");
           }}